Update styles.
This commit updates the style conventions used by magi.
We use to use the yapf autoformatter with Google style 2 space indents.

This can make it difficult to other people to collaborate with us since
more people are more farmiliar with PEP8 and are used to 4 spaces.

Therefore, in order for others to set up things more easily and provide
a style where people can easily copy the files and adapt the code
without re-formatting. We decide to start using 4 spaces for indents.
In addition, there is no reason to continue using yapf so we switch
to black since, in our experience, it offers better formatting
in some cases.

To accomodate the increase in indents, we have bumped the line length
limit to 88. Sometimes black produces lines that are slightly
above 88, we have decided to increase the line limit in flake8 and
pylint to 100 to be more relaxed on line limits.

The rest of the coding style should still follow the Google
style guide, which is PEP8 compatible. We additionally enforce
import order using pep8-import-order to be consistent with
the Google style.