Add PETS (#16)
Add PETS, a model-based RL algorithm based on NN ensembles.

This commit includes a simple implementation of PETS.
Features implemented are the TSInf propagation method.

The other methods (TS0, MM) should be straightforward to incorporate
by changing `propagate()` in EnsembleModel.

The Ensemble model keeps track of the normalizer, which
is used to standardize the observations. This is found to stabilize training
and improve performance. Especially on harder tasks such as HalfCheetah.

Implementation is benchmarked on HalfCheetah, which can obtain similar
performance compared to the ones in mbrl-lib. It works quite well on CartPole
and can obtain optimal performance on CartPole with <10 trials.

Additional notes
* model unrolling and CEM is implemented with lax.scan and lax.while_loop
to speed up compilation.
* The details are set up to match mbrl-lib as closely as we can - which deviates
from the original PETS implementation. See 
https://github.com/facebookresearch/mbrl-lib/issues/99 
for a details summary of the differences. I have not ablated on these differences in
details.