Add AWAC agent implementation
* Add the Advantage Weighted Actor Critic (AWAC) algorithm
from (Nair et al., 2020)[1].

* Include two examples. The first is used as a offline RL algorithm
on the D4RL dataset. The second example demonstrates how to perform
pre-training with online finetuning on D4RL tasks.

Notes:
- This implementation is implemented with reference to

https://github.com/ikostrikov/jaxrl/blob/main/jaxrl/agents/awac/awac_learner.py

and is a port from the Flax implementation to use Haiku + Acme.

- Another source of reference would be

https://www.github.com/rail-berkeley/rlkit/tree/master/rlkit

which is the "official" implementation, this version however includes
a lot of flags to switch between different algorithms and is difficult
to understand what is happening.

- AWAC is very similar to the CRR algorithm, with differences
in how to normalize the advantages and the base RL off-policy
algorithm.

References:
[1]: Nair, A., Dalal, M., Gupta, A., & Levine, S. (2020). Accelerating
     online reinforcement learning with offline datasets.
     arXiv preprint arXiv:2006.09359.
     https://arxiv.org/abs/2006.09359