feat(drq-v2): Add DrQ-v2 agent
Add a JAX implementation of DrQ-v2 from
---

> Yarats, D., Fergus, R., Lazaric, A., & Pinto, L. (2021).
_Mastering visual continuous control: Improved data-augmented
reinforcement learning_. arXiv preprint
[arXiv:2107.09645](https://arxiv.org/abs/2107.09645).

The official PyTorch implementation can be found at

    https://github.com/facebookresearch/drqv2

As with other agents, it uses Reverb as the backend for replay.
However, this may be problematic if we are interested
in reproducing the results from the paper, which uses a 1M
replay since Reverb does not have mechanisms for storing
part of the replay on disk. It is worth investigating the
actual memory usage for using a 1M replay. It might be
feasible to still use this implementation since Reverb
also does compression under the hood.

There are some remaining TODOs, which can be found in
magi/agents/drq_v2/README.md. Nevertheless, this implementation
should match the original implementation in most of the details.

Refactor DrQ implementation
---
This commit also refactors the existing DrQ agent.
Some types were simplified.

Closes #52.