(style): Revert to Google Python style
After some careful consideration, we will revert to
Google's Python style for OSS projects.
The majority of this change consists of
using single-quote for strings. Two-space
indentation. And switching from black to yapf for
formatting.

The main reason for this change is to allow us to
fork components from Acme (or more generally, the
the few DeepMind JAX ecosystem libraries) more transparently.
Having a consistent style as these libraries allow us
to fork an individual module/package and make the necessary
changes, without having to reformat the entire file.
This should give us a cleaner diff when we want to compare
with the version upstream. This can also potentially
make contributing back to these libraries easier.